+++ src/czml3/examples/__init__.py
@@ -1,0 +1,3 @@
+from .simple import simple
+
+__all__ = ["simple"]
